refactor: consolidate prompt utilities under src/utils/prompt
Move scattered prompt-related utilities from src/agent/utils/ to a
centralized location at src/utils/prompt/. This improves code
organization and makes prompt utilities more discoverable.

Files consolidated:
- promptUtils.ts: Core utilities (XML formatting, rendering, file ops)
- promptHelpers.ts: System prompt composition with rules
- PromptBuilder.ts: Multi-round prompt construction class
- toolUsePrompt.ts: Tool-use instruction constants

The agent/utils/index.ts now re-exports from the new location for
backwards compatibility through the @agent/utils alias.

Also fixes unrelated pre-existing TypeScript error in
BaseViewContentProvider.ts (readonly array parameter type).
