+++ src/agent/utils/index.ts
@@ -1,9 +1,6 @@
 export * from './messageSkeletonUtils';
 export * from './userVars';
 export * from './mediaTypes';
-export * from './promptUtils';
-export * from './promptHelpers';
-export * from './PromptBuilder';
 export * from './outputFileUtils';
 export * from './priceUtils';
 export * from './text';
@@ -11,3 +8,6 @@ export * from './UsageMonitor';
 export * from './debugMessageSaver';
 export * from './continuationMessage';
 export * from './mergeFileUtils';
+
+// Re-export prompt utilities from consolidated location
+export * from '@utils/prompt';
